+++ siteManagerDashboard/index.js
@@ -5,6 +5,8 @@ import { renderParticipantDetails } from './participantDetails.js';
 import { renderParticipantSummary } from './participantSummary.js'
 import { internalNavigatorHandler, humanReadableY } from './utils.js'
 import fieldMapping from './fieldToConceptIdMapping.js';
+import { nameToKeyObj } from './siteKeysToName.js';
+import { renderAllCharts } from './participantChartsRender.js'
 
 let saveFlag = false;
 let counter = 0;
@@ -106,74 +108,56 @@ const renderDashboard = async () => {
     }
 }
 
-// const reRenderDashboard = async (siteKey) => {
-//     if (siteKey) {
-//         animation(true);
-//         const isAuthorized = await authorize(siteKey);
-//         if (isAuthorized && isAuthorized.code === 200) {
+const renderSiteKeyList = () => {
+    let template = ``;
+    template += `
+            <div style="margin-top:10px; padding:15px;" class="dropdown">
+                <button class="btn btn-secondary dropdown-toggle" type="button"  data-toggle="dropdown" aria-haspopup="true" aria-expanded="false">
+                Change Site Preference 
+                </button>
+            <div class="dropdown-menu" id="dropdownMenuButton" aria-labelledby="dropdownMenuButton">
+                <a class="dropdown-item" data-siteKey="hfHealth" id="hfHealth">Henry Ford Health Systems</a>
+                <a class="dropdown-item" data-siteKey="hPartners" id="hPartners">Health Partners</a>
+                <a class="dropdown-item" data-siteKey="kpGA" id="kpGA">KP GA</a>
+                <a class="dropdown-item" data-siteKey="kpHI" id="kpHI">KP HI</a>
+                <a class="dropdown-item" data-siteKey="kpNW" id="kpNW">KP NW</a>
+                <a class="dropdown-item" data-siteKey="kpCO" id="kpCO">KP CO</a>
+                <a class="dropdown-item" data-siteKey="maClinic" id="maClinic">Marshfield Clinic</a>
+                <a class="dropdown-item" data-siteKey="nci" id="nci">NCI</a>
+                <a class="dropdown-item" data-siteKey="snfrdHealth" id="snfrdHealth">Sanford Health</a>
+                <a class="dropdown-item" data-siteKey="uChiM" id="uChiM">UofC Medicine</a>
+            </div>
+        </div>
+            `
+    return template;
+}
 
-//             document.getElementById('navBarLinks').innerHTML = dashboardNavBarLinks();
-//             removeActiveClass('nav-link', 'active');
-//             document.getElementById('dashboardBtn').classList.add('active');
-//             mainContent.innerHTML = '';
+const dropdownTrigger = (filterWorkflowResults, participantsGenderMetric, participantsRaceMetric, participantsAgeMetric, filterVerificationResults, recruitsCountResults) => {
 
-//             renderCharts(siteKey);
-//         }
-//         internalNavigatorHandler(counter); // function call to prevent internal navigation when there's unsaved changes
-//         if (isAuthorized.code === 401) {
-//             clearLocalStroage();
-//         }
-//     } else {
-//         animation(false);
-//         window.location.hash = '#';
-//     }
-// }
+    const dropdownMenuButton = document.getElementById('dropdownMenuButton')
+    if (dropdownMenuButton) {
+        dropdownMenuButton.addEventListener('click', (e) => {
+            const t = getDataAttributes(e.target)
+            console.log('t', t)
+           reRenderDashboard(t.sitekey, filterWorkflowResults, participantsGenderMetric, participantsRaceMetric, participantsAgeMetric, filterVerificationResults, recruitsCountResults);
+        })
+    }
+}
 
-// const renderSiteKeyList = (siteKey) => {
-//     let template = ``;
-//     if (siteKey === fieldMapping['nci'] || fieldMapping['norc']) {
-  
-//     template += `
-//             <div style="margin-top:10px; padding:15px;" class="dropdown">
-//                 <button class="btn btn-secondary dropdown-toggle" type="button"  data-toggle="dropdown" aria-haspopup="true" aria-expanded="false">
-//                 Change Site Preference 
-//                 </button>
-//             <div class="dropdown-menu" id="dropdownMenuButton" aria-labelledby="dropdownMenuButton">
-//                 <a class="dropdown-item" data-siteKey="kpGA" id="kpGA">KP GA</a>
-//                 <a class="dropdown-item" data-siteKey="kpHI" id="kpHI">KP HI</a>
-//                 <a class="dropdown-item" data-siteKey="nci" id="nci">NCI</a>
-//                 <a class="dropdown-item" data-siteKey="kpCO" id="kpCO">KP CO</a>
-//                 <a class="dropdown-item" data-siteKey="maClinic" id="maClinic">Marshfield Clinic</a>
-//                 <a class="dropdown-item" data-siteKey="hfHealth" id="hfHealth">Henry Ford Health Systems</a>
-//             </div>
-//         </div>
-//             `}
-//     return template;
-// }
+const getDataAttributes = (el) => {
+    let data = {};
+    [].forEach.call(el.attributes, function(attr) {
+        if (/^data-/.test(attr.name)) {
+            var camelCaseName = attr.name.substr(5).replace(/-(.)/g, function ($0, $1) {
+                return $1.toUpperCase();
+            });
+            data[camelCaseName] = attr.value;
+        }
+    });
+    return data;
+}
 
-// const dropdownTrigger = () => {
 
-//     const dropdownMenuButton = document.getElementById('dropdownMenuButton')
-//     if (dropdownMenuButton) {
-//         dropdownMenuButton.addEventListener('click', (e) => {
-//             const t = getDataAttributes(e.target)
-//             reRenderDashboard(fieldMapping[t.sitekey]);
-//         })
-//     }
-// }
-
-// const getDataAttributes = (el) => {
-//     let data = {};
-//     [].forEach.call(el.attributes, function(attr) {
-//         if (/^data-/.test(attr.name)) {
-//             var camelCaseName = attr.name.substr(5).replace(/-(.)/g, function ($0, $1) {
-//                 return $1.toUpperCase();
-//             });
-//             data[camelCaseName] = attr.value;
-//         }
-//     });
-//     return data;
-// }
 
 const fetchData = async (siteKey, type) => {
     if (!checkSession()) {
@@ -198,6 +182,22 @@ const fetchStats = async (siteKey, type) => {
     }
     else {
         const response = await fetch(`https://us-central1-nih-nci-dceg-connect-dev.cloudfunctions.net/stats?type=${type}`, {
+            method: 'GET',
+            headers: {
+                Authorization: "Bearer " + siteKey
+            }
+        });
+        return response.json();
+    }
+}
+
+const localFetchStats = async (siteKey, type) => {
+    if (!checkSession()) {
+        alert('Session expired!');
+        clearLocalStroage();
+    }
+    else {
+        const response = await fetch(`http://localhost:5001/nih-nci-dceg-connect-dev/us-central1/stats?type=${type}`, {
             method: 'GET',
             headers: {
                 Authorization: "Bearer " + siteKey
@@ -273,7 +273,7 @@ const renderCharts = async (siteKey) => {
 
     const participantsGenderMetric = await fetchStats(siteKey, 'sex');
     const participantsRaceMetric = await fetchStats(siteKey, 'race');
-    const participantsAgeMetric = await fetchStats(siteKey, 'age');
+    const participantsAgeMetric = await localFetchStats(siteKey, 'age');
 
     const activeVerificationStatus = filterVerification(filterVerificationResults.stats, 'active')
     const passiveVerificationStatus = filterVerification(filterVerificationResults.stats, 'passive')
@@ -282,175 +282,23 @@ const renderCharts = async (siteKey) => {
 
     const recruitsCount = filterRecruits(recruitsCountResults.stats)
 
-    if (stats.code === 200) {
-        const siteSelectionRow = document.createElement('div');
-        siteSelectionRow.classList = ['row'];
-        siteSelectionRow.id = 'siteSelection';
-        if (stats.data.length > 1) {
-            //mainContent.appendChild(siteSelectionRow);
-            //renderSiteSelection(stats.data);
-        }
-        // const rowHeader = document.createElement('div');
-        // rowHeader.classList = ['rowHeader'];
-        // rowHeader.innerHTML = renderSiteKeyList(siteKey)
-        // mainContent.appendChild(rowHeader);
-        // dropdownTrigger();
+    const objSiteCode = recruitsCount[0] && recruitsCount[0].siteCode
+    const siteSelectionRow = document.createElement('div');
+    siteSelectionRow.classList = ['row'];
+    siteSelectionRow.id = 'siteSelection';
+    if (recruitsCountResults.code === 200) {    
+        recruitsCountResults && recruitsCountResults.stats.forEach((i, index) => {
+              if (index !== 0 && (objSiteCode !== i.siteCode)) {
+                siteSelectionRow.innerHTML = renderSiteKeyList(siteKey)
+                mainContent.appendChild(siteSelectionRow);
+                dropdownTrigger(filterWorkflowResults.stats, participantsGenderMetric, participantsRaceMetric, participantsAgeMetric, 
+                    filterVerificationResults.stats, recruitsCountResults.stats);
+              }
+        }) 
 
-        const row = document.createElement('div');
-        row.classList = ['row'];
-
-        let funnelChart = document.createElement('div');
-        funnelChart.classList = ['col-lg-4 charts'];
-
-        let subFunnelChart = document.createElement('div');
-        subFunnelChart.classList = ['col-lg-12 sub-div-shadow viz-div'];
-        subFunnelChart.innerHTML = renderLabel(recruitsCount.activeCount, 'Active');
-        subFunnelChart.setAttribute('id', 'funnelChart');
-        funnelChart.appendChild(subFunnelChart);
-
-        let barChart = document.createElement('div');
-        barChart.classList = ['col-lg-4 charts'];
-
-        let subBarChart = document.createElement('div');
-        subBarChart.classList = ['col-lg-12 sub-div-shadow viz-div'];
-        subBarChart.innerHTML = renderLabel(recruitsCount.activeCount, 'Active');
-        subBarChart.setAttribute('id', 'barChart');
-        barChart.appendChild(subBarChart);
-
-        let activeOptouts = document.createElement('div');
-        activeOptouts.classList = ['col-lg-4 charts'];
-
-        let subactiveOptouts = document.createElement('div');
-        subactiveOptouts.classList = ['col-lg-12 sub-div-shadow viz-div'];
-    //    activeOptouts.innerHTML = renderLabel(recruitsCount.activeCount, 'Active')
-        subactiveOptouts.setAttribute('id', 'activeOptouts');
-        activeOptouts.appendChild(subactiveOptouts);
-
-        row.appendChild(funnelChart);
-        row.appendChild(barChart);
-        row.appendChild(activeOptouts);
-
-        mainContent.appendChild(row); // active 
-
-        const row1 = document.createElement('div');
-        row1.classList = ['row'];
-
-        let funnelChart1 = document.createElement('div');
-        funnelChart1.classList = ['col-lg-4 charts'];
-
-        let subFunnelChart1 = document.createElement('div');
-        subFunnelChart1.classList = ['col-lg-12 viz-div sub-div-shadow'];
-        subFunnelChart1.innerHTML = renderLabel(recruitsCount.passiveCount, 'Passive');
-        subFunnelChart1.setAttribute('id', 'passiveFunnelChart');
-        funnelChart1.appendChild(subFunnelChart1);
-
-        let barChart1 = document.createElement('div');
-        barChart1.classList = ['col-lg-4 charts'];
-
-        let subBarChart1 = document.createElement('div');
-        subBarChart1.classList = ['col-lg-12 viz-div sub-div-shadow'];
-        subBarChart1.innerHTML = renderLabel(recruitsCount.passiveCount, 'Passive');
-        subBarChart1.setAttribute('id', 'passiveBarChart');
-        barChart1.appendChild(subBarChart1);
-
-        let activeCounts1 = document.createElement('div');
-        activeCounts1.classList = ['col-lg-4 charts'];
-        let subActiveCounts1 = document.createElement('div');
-        subActiveCounts1.classList = ['col-lg-12 viz-div sub-div-shadow'];
-       // subActiveCounts1.innerHTML = renderLabel(recruitsCount.passiveCount, 'Passive');
-        subActiveCounts1.setAttribute('id', 'passiveCounts');
-        activeCounts1.appendChild(subActiveCounts1);
-
-
-        row1.appendChild(funnelChart1);
-        row1.appendChild(barChart1);
-        row1.appendChild(activeCounts1);
-
-        mainContent.appendChild(row1); // passive
-
-        const row2 = document.createElement('div');
-        row2.classList = ['row'];
-
-        let funnelChart2 = document.createElement('div');
-        funnelChart2.classList = ['col-lg-4 charts'];
-
-        let subFunnelChart2 = document.createElement('div');
-        subFunnelChart2.classList = ['col-lg-12 viz-div sub-div-shadow'];
-        subFunnelChart2.innerHTML = renderLabel(recruitsCount.activeCount + recruitsCount.passiveCount, 'Total');
-        subFunnelChart2.setAttribute('id', 'totalFunnelChart');
-        funnelChart2.appendChild(subFunnelChart2);
-        row2.appendChild(funnelChart2)
-
-        let barChart2 = document.createElement('div');
-        barChart2.classList = ['col-lg-4 charts'];
-
-        let subBarChart2 = document.createElement('div');
-        subBarChart2.classList = ['col-lg-12 viz-div sub-div-shadow'];
-        subBarChart2.innerHTML = renderLabel(recruitsCount.activeCount + recruitsCount.passiveCount, 'Total');;
-        subBarChart2.setAttribute('id', 'totalBarChart');
-        barChart2.appendChild(subBarChart2);
-        row2.appendChild(barChart2);
-
-        let totalCounts = document.createElement('div');
-        totalCounts.classList = ['col-lg-4 charts'];
-
-        let subTotalCounts = document.createElement('div');
-        subTotalCounts.classList = ['col-lg-12 sub-div-shadow viz-div'];
-      //  subTotalCounts.innerHTML = renderLabel(recruitsCount.activeCount + recruitsCount.passiveCount, 'Total');;
-        subTotalCounts.setAttribute('id', 'totalCounts');
-        totalCounts.appendChild(subTotalCounts);
-        row2.appendChild(totalCounts);
-
-        mainContent.appendChild(row2); // total
-
-        const row3 = document.createElement('div');
-        row3.classList = ['row'];
-
-        let stackedBarChart = document.createElement('div');
-        stackedBarChart.classList = ['col-lg-4 charts'];
-
-        let subStackedBarChart = document.createElement('div');
-        subStackedBarChart.classList = ['col-lg-12 viz-div sub-div-shadow'];
-        subStackedBarChart.setAttribute('id', 'metrics');
-        stackedBarChart.appendChild(subStackedBarChart);
-        row3.appendChild(stackedBarChart);
-
-        let pieChart = document.createElement('div');
-        pieChart.classList = ['col-lg-4 charts'];
-
-        let subPieChart = document.createElement('div');
-        subPieChart.classList = ['col-lg-12 viz-div sub-div-shadow'];
-        subPieChart.setAttribute('id', 'activeVerificationStatus');
-        pieChart.appendChild(subPieChart);
-        row3.appendChild(pieChart);
-
-        let pieChart1 = document.createElement('div');
-        pieChart1.classList = ['col-lg-4 charts'];
-
-        let subPieChart1 = document.createElement('div');
-        subPieChart1.classList = ['col-lg-12 viz-div sub-div-shadow'];
-        subPieChart1.setAttribute('id', 'passiveVerificationStatus');
-        pieChart1.appendChild(subPieChart1);
-        row3.appendChild(pieChart1);
-
-        mainContent.appendChild(row3); // Misc.
-
-
-
-        renderAllMetricCharts(participantsGenderMetric, participantsRaceMetric, participantsAgeMetric)
-
-        renderActiveFunnelChart(activeRecruitsFunnel, 'funnelChart')
-        renderActiveBarChart(activeCurrentWorkflow, 'barChart');
-        renderActiveOptouts('activeOptouts');
-
-        renderPassiveFunnelChart(passiveRecruitsFunnel, 'passiveFunnelChart');
-        renderPassiveBarChart(passiveCurrentWorkflow, 'passiveBarChart');
-
-        renderTotalFunnelChart(totalRecruitsFunnel, 'totalFunnelChart');
-        renderTotalCurrentWorkflow(totalCurrentWorkflow, 'totalBarChart');
-
-        renderActiveVerificationStatus(activeVerificationStatus, denominatorVerificationStatus, 'activeVerificationStatus');
-        renderPassiveVerificationStatus(passiveVerificationStatus, denominatorVerificationStatus, 'passiveVerificationStatus');
+    renderAllCharts(activeRecruitsFunnel, passiveRecruitsFunnel, totalRecruitsFunnel, activeCurrentWorkflow, passiveCurrentWorkflow, totalCurrentWorkflow, 
+        participantsGenderMetric, participantsRaceMetric, participantsAgeMetric, activeVerificationStatus, passiveVerificationStatus, 
+        denominatorVerificationStatus, recruitsCount )
 
         animation(false);
     }
@@ -645,636 +493,73 @@ const filterRecruits = (data) => {
     return currentObj;
 }
 
-const renderLabel = (count, recruitType) => {
-    let template = ``;
-    (recruitType === 'Active') ?
-        (template += `<span class="badge bg-primary" style="color:white"> ${recruitType} Recruits: ${count}</span>`)
-        :
-    (recruitType === 'Passive') ?
-        (template += `<span class="badge bg-primary" style="color:white"> ${recruitType} Recruits: ${count}</span>`)
-        :
-    (recruitType === 'Total') ?
-        (template += `<span class="badge bg-primary" style="color:white"> ${recruitType} Recruits: ${count}</span>`)
-        :
-        ''
 
-    return template;
+const reRenderDashboard = async (siteKey, filterWorkflowResults, participantsGenderMetric, participantsRaceMetric, 
+    participantsAgeMetric, filterVerificationResults, recruitsCountResults) => {
 
+console.log('sk',participantsGenderMetric )
+const siteKeyFilter = nameToKeyObj[siteKey]
+
+// filterDatabySiteCode(siteKeyFilter, filterWorkflowResults)
+let resultWorkflow = []
+filterDatabySiteCode(resultWorkflow, filterWorkflowResults, siteKeyFilter)
+
+let resultGender = []
+participantsGenderMetric.stats.filter(i => {
+    if(i.siteCode === siteKeyFilter) {
+        resultGender.push(i)
+}});
+let resultRace = []
+participantsRaceMetric.stats.filter(i => {
+    if(i.siteCode === siteKeyFilter) {
+        resultRace.push(i)
+}});
+let resultAge = []
+participantsAgeMetric.stats.filter(i => {
+    if(i.siteCode === siteKeyFilter) {
+        resultAge.push(i)
+}});
+let resultVerification = []
+filterVerificationResults.filter(i => {
+    if(i.siteCode === siteKeyFilter) {
+        resultVerification.push(i)
+}});
+let resultRecruitsCount = []
+recruitsCountResults.filter(i => {
+    if(i.siteCode === siteKeyFilter) {
+        resultRecruitsCount.push(i)
+}});
+
+mainContent.innerHTML = '';
+
+const activeRecruitsFunnel = filterRecruitsFunnel(resultWorkflow, 'active')
+const passiveRecruitsFunnel = filterRecruitsFunnel(resultWorkflow, 'passive')
+const totalRecruitsFunnel = filterTotalRecruitsFunnel(resultWorkflow)
+
+const activeCurrentWorkflow = filterCurrentWorkflow(resultWorkflow, 'active')
+const passiveCurrentWorkflow = filterCurrentWorkflow(resultWorkflow, 'passive')
+const totalCurrentWorkflow = filterTotalCurrentWorkflow(resultWorkflow)
+
+const activeVerificationStatus = filterVerification(resultVerification, 'active')
+const passiveVerificationStatus = filterVerification(resultVerification, 'passive')
+const denominatorVerificationStatus = filterDenominatorVerificationStatus(resultVerification)
+
+//renderAllMetricCharts(participantsGenderMetric, participantsRaceMetric, participantsAgeMetric)
+const recruitsCount = filterRecruits(resultRecruitsCount)
+
+const siteSelectionRow = document.createElement('div');
+siteSelectionRow.classList = ['row'];
+siteSelectionRow.id = 'siteSelection';
+
+siteSelectionRow.innerHTML = renderSiteKeyList(siteKey)
+mainContent.appendChild(siteSelectionRow);
+dropdownTrigger(filterWorkflowResults, participantsGenderMetric, participantsRaceMetric, 
+    participantsAgeMetric, filterVerificationResults, recruitsCountResults)
+
+renderAllCharts(activeRecruitsFunnel, passiveRecruitsFunnel, totalRecruitsFunnel, activeCurrentWorkflow, passiveCurrentWorkflow, totalCurrentWorkflow, resultGender, resultRace, resultAge, activeVerificationStatus, passiveVerificationStatus, denominatorVerificationStatus, recruitsCount )
+
+ animation(false);
 }
-
-const renderActiveFunnelChart = (activeRecruitsFunnel, id) => {
-    const signIn =  activeRecruitsFunnel.signedIn
-    const consent = activeRecruitsFunnel.consented;
-    const userProfile =  activeRecruitsFunnel.submittedProfile
-    const verification = activeRecruitsFunnel.verification
-    const verified = activeRecruitsFunnel.verified
-    
-        const data = [{
-            x: [signIn, consent, userProfile, verification, verified],
-            y: ['Signed In', 'Consented', 'Submitted', 'Verification', 'Verified'],
-            type: 'funnel',
-            textinfo: 'hello ${23}', 
-            marker: {
-                color: ["#0C1368", "#242C8F", "#525DE9", '#008ECC', '#6593F5']
-            }
-        }];
-    
-        const layout = {
-            paper_bgcolor: 'rgba(0,0,0,0)',
-            plot_bgcolor: 'rgba(0,0,0,0)',
-            title: `Cumulative Status`
-        };
-    
-        Plotly.newPlot(id, data, layout, { responsive: true, displayModeBar: false });
-    }
-
-
-const renderActiveBarChart = (activeCurrentWorkflow, id) => {
-    const notSignedIn = activeCurrentWorkflow.notSignedIn;
-    const signedInNotConsented = activeCurrentWorkflow.signedIn;
-    const consentedNotSubmitted = activeCurrentWorkflow.consented;
-    const submittedVerificationNotCompleted = activeCurrentWorkflow.submittedProfile;
-    const verificationCompleted = activeCurrentWorkflow.verification;
-    const trace1 = {
-        x: ['Never Signed In', 'Signed In, No Consent', 'Consented, No Profile', 'Profile, Verified Not Complete', 'Verification Complete'],
-        y: [notSignedIn, signedInNotConsented, consentedNotSubmitted, submittedVerificationNotCompleted, verificationCompleted],
-        name: 'Completed',
-        type: 'bar'
-    };
-
-    const data = [trace1];
-
-    const layout = {
-        barmode: 'stack',
-        showlegend: false,
-        paper_bgcolor: 'rgba(0,0,0,0)',
-        plot_bgcolor: 'rgba(0,0,0,0)',
-        yaxis: {
-            automargin: true,
-            fixedrange: true
-        },
-        xaxis: {
-            automargin: true,
-            fixedrange: true
-        },
-       // colorway: ['#7f7fcc', '#0C1368', '#525DE9', '#008ECC'],
-        title: 'Current Status in Workflow'
-    };
-
-    Plotly.newPlot(id, data, layout, { responsive: true, displayModeBar: false });
-}
-
-const renderPassiveFunnelChart = (passiveRecruitsFunnel, id) => {
-    
-    const signIn =  passiveRecruitsFunnel.signedIn
-    const consent = passiveRecruitsFunnel.consented;
-    const userProfile =  passiveRecruitsFunnel.submittedProfile
-    const verification = passiveRecruitsFunnel.verification
-    const verified = passiveRecruitsFunnel.verified
-
-        const data = [{
-            x: [signIn, consent, userProfile, verification, verified],
-            y: ['Signed In', 'Consented', 'Submitted', 'Verification', 'Verified'],
-            type: 'funnel',
-            marker: {
-                color: ["#0C1368", "#242C8F", "#525DE9", '#008ECC', '#6593F5']
-            }
-        }];
-    
-        const layout = {
-            paper_bgcolor: 'rgba(0,0,0,0)',
-            plot_bgcolor: 'rgba(0,0,0,0)'
-        };
-    
-        Plotly.newPlot(id, data, layout, { responsive: true, displayModeBar: false });
-};
-
-const renderActiveOptouts = (id) => {
-    let template = ``;
-    template += `
-        <span style="text-align: center;"><h5>Opt Outs</h5></span>`
-    document.getElementById(id).innerHTML = template
-}
-
-const renderPassiveBarChart = (passiveCurrentWorkflow, id) => {
-    const notSignedIn = passiveCurrentWorkflow.notSignedIn;
-    const signedInNotConsented = passiveCurrentWorkflow.signedIn;
-    const consentedNotSubmitted = passiveCurrentWorkflow.consented;
-    const submittedVerificationNotCompleted = passiveCurrentWorkflow.submittedProfile;
-    const verificationCompleted = passiveCurrentWorkflow.verification;
-    const trace1 = {
-        x: ['Not signed in ', 'Signed in not consent ', ' Consented User Profile not Submitted', 'Submitted Verification Not Completed', 'Verification'],
-        y: [notSignedIn, signedInNotConsented, consentedNotSubmitted, submittedVerificationNotCompleted, verificationCompleted],
-        name: 'Completed',
-        type: 'bar'
-    };
-
-    const data = [trace1];
-
-    const layout = {
-        barmode: 'stack',
-        showlegend: false,
-        paper_bgcolor: 'rgba(0,0,0,0)',
-        plot_bgcolor: 'rgba(0,0,0,0)',
-        yaxis: {
-            automargin: true,
-            fixedrange: true
-        },
-        xaxis: {
-            automargin: true,
-            fixedrange: true
-        },
-       // colorway: ['#7f7fcc', '#0C1368', '#525DE9', '#008ECC'],
-    };
-
-    Plotly.newPlot(id, data, layout, { responsive: true, displayModeBar: false });
-
-}
-
-const renderTotalFunnelChart = (totalRecruitsFunnel, id) => {
-
-    const signIn =  totalRecruitsFunnel.signedIn
-    const consent = totalRecruitsFunnel.consented;
-    const userProfile =  totalRecruitsFunnel.submittedProfile
-    const verified = totalRecruitsFunnel.verified
-    const verification = totalRecruitsFunnel.verification
-    
-        const data = [{
-            x: [signIn, consent, userProfile, verification, verified],
-            y: ['Signed In', 'Consented', 'Submitted', 'Verification', 'Verified'],
-            type: 'funnel',
-            marker: {
-                color: ["#0C1368", "#242C8F", "#525DE9", '#008ECC', '#6593F5']
-            }
-        }];
-    
-        const layout = {
-            paper_bgcolor: 'rgba(0,0,0,0)',
-            plot_bgcolor: 'rgba(0,0,0,0)',
-        };
-    
-        Plotly.newPlot(id, data, layout, { responsive: true, displayModeBar: false });
-};
-
-
-const renderTotalCurrentWorkflow = (totalCurrentWorkflow, id) => {
-    const notSignedIn = totalCurrentWorkflow.notSignedIn;
-    const signedInNotConsented = totalCurrentWorkflow.signedIn;
-    const consentedNotSubmitted = totalCurrentWorkflow.consented;
-    const submittedVerificationNotCompleted = totalCurrentWorkflow.submittedProfile;
-    const verificationCompleted = totalCurrentWorkflow.verification;
-    const trace1 = {
-        x: ['Not signed in ', 'Signed in not consent ', ' Consented User Profile not Submitted', 'Submitted Verification Not Completed', 'Verification'],
-        y: [notSignedIn, signedInNotConsented, consentedNotSubmitted, submittedVerificationNotCompleted, verificationCompleted],
-        name: 'Completed',
-        type: 'bar'
-    };
-
-    const data = [trace1];
-
-    const layout = {
-        barmode: 'stack',
-        showlegend: false,
-        paper_bgcolor: 'rgba(0,0,0,0)',
-        plot_bgcolor: 'rgba(0,0,0,0)',
-        yaxis: {
-            automargin: true,
-            fixedrange: true
-        },
-        xaxis: {
-            automargin: true,
-            fixedrange: true
-        },
-       // colorway: ['#7f7fcc', '#0C1368', '#525DE9', '#008ECC'],
-    };
-
-    Plotly.newPlot(id, data, layout, { responsive: true, displayModeBar: false });
-
-}
-
-const renderActiveVerificationStatus = (activeVerificationStatus, denominatorVerificationStatus, id) => {
-    const notYetVerified =  activeVerificationStatus.notYetVerified 
-                            && ((activeVerificationStatus.notYetVerified)/(denominatorVerificationStatus.activeDenominator)*100).toFixed(2);
-    const verified = activeVerificationStatus.verified 
-                        && ((activeVerificationStatus.verified)/(denominatorVerificationStatus.activeDenominator)*100).toFixed(2);
-    const cannotBeVerified =  activeVerificationStatus.cannotBeVerified 
-                        && ((activeVerificationStatus.cannotBeVerified)/(denominatorVerificationStatus.activeDenominator)*100).toFixed(2);
-    const duplicate = activeVerificationStatus.duplicate 
-                        && ((activeVerificationStatus.duplicate)/(denominatorVerificationStatus.activeDenominator)*100).toFixed(2);
-    const outreachTimedOut = activeVerificationStatus.outreachTimedOut 
-                        && ((activeVerificationStatus.outreachTimedOut)/(denominatorVerificationStatus.activeDenominator)*100).toFixed(2);
-
-    var data = [{
-        values: [notYetVerified, verified, cannotBeVerified, duplicate, outreachTimedOut],
-        labels: [ 'Not Verified', 'Verified', 'Cannot be Verified','Duplicate', 'Outreach Maxed Out'],
-        type: 'pie'
-      }];
-      
-      const layout = {
-
-        showlegend: true,
-        paper_bgcolor: 'rgba(0,0,0,0)',
-        plot_bgcolor: 'rgba(0,0,0,0)',
-       // colorway: ['#7f7fcc', '#0C1368', '#525DE9', '#008ECC'],
-        title: 'Active Recruits Verification Status'
-    };
-      
-
-  
-      Plotly.newPlot(id, data, layout, { responsive: true, displayModeBar: false });
-  }
-
-  const renderPassiveVerificationStatus = (passiveVerificationStatus, denominatorVerificationStatus, id) => {
-    const notYetVerified =  passiveVerificationStatus.notYetVerified 
-                            && ((passiveVerificationStatus.notYetVerified)/(denominatorVerificationStatus.passiveDenominator)*100).toFixed(2);
-    const verified = passiveVerificationStatus.verified 
-                        && ((passiveVerificationStatus.verified)/(denominatorVerificationStatus.passiveDenominator)*100).toFixed(2);
-    const cannotBeVerified =  passiveVerificationStatus.cannotBeVerified 
-                        && ((passiveVerificationStatus.cannotBeVerified)/(denominatorVerificationStatus.passiveDenominator)*100).toFixed(2);
-    const duplicate = passiveVerificationStatus.duplicate 
-                        && ((passiveVerificationStatus.duplicate)/(denominatorVerificationStatus.passiveDenominator)*100).toFixed(2);
-     const outreachTimedOut = passiveVerificationStatus.outreachTimedOut 
-                        && ((passiveVerificationStatus.outreachTimedOut)/(denominatorVerificationStatus.passiveDenominator)*100).toFixed(2);
-    var data = [{
-        values: [notYetVerified, verified, cannotBeVerified, duplicate, outreachTimedOut],
-        labels: [ 'Not Verified', 'Verified', 'Cannot be Verified','Duplicate', 'Outreach Maxed Out'],
-        type: 'pie'
-      }];
-
-      const layout = {
-        showlegend: true,
-        paper_bgcolor: 'rgba(0,0,0,0)',
-        plot_bgcolor: 'rgba(0,0,0,0)',
-       // colorway: ['#7f7fcc', '#0C1368', '#525DE9', '#008ECC'],
-        title: 'Passive Recruits Verification Status'
-    };
-      
-      Plotly.newPlot(id, data, layout, { responsive: true, displayModeBar: false });
-  }
-
-  const renderAllMetricCharts = (participantsGenderMetric, participantsRaceMetric, participantsAgeMetric) => {
-     
-    let genderObject = {female: 0, male: 0, intersex: 0, unknown: 0}
-    filterGenderMetrics(participantsGenderMetric.stats, genderObject)
-    let raceObject = {americanIndian: 0, asian: 0, africanAmerican: 0, latino: 0, nativeHawaiian: 0, middleEastern: 0, white: 0, none: 0, other: 0}
-    filterRaceMetrics(participantsRaceMetric.stats, raceObject)
-    console.log('participantsGenderMetric', participantsAgeMetric)
-    let ageObject = {'40-45': 0, '46-50': 0, '51-55': 0, '56-60': 0, '61-65': 0}
-    filterAgeMetrics(participantsAgeMetric.stats, ageObject)
-
-    renderStackBarChart(genderObject, raceObject, ageObject, 'metrics')
-   
-}
-
-const filterGenderMetrics = (participantsGenderMetrics, genderObject) => {
-        participantsGenderMetrics.forEach( i => {
-        (parseInt(i.sex) === fieldMapping['male']) ?
-            genderObject['male'] = parseInt(i.sexCount)
-        :
-        (parseInt(i.sex) === fieldMapping['female']) ?
-            genderObject['female'] = parseInt(i.sexCount)
-        :
-        (parseInt(i.sex) === fieldMapping['intersex']) ?
-            genderObject['intersex'] = parseInt(i.sexCount)
-        :
-            genderObject['unknown'] = parseInt(i.sexCount)
-        return genderObject;
-        }
-    )}
-
-
-const filterRaceMetrics = (participantsRaceMetrics, raceObject) => {
-    participantsRaceMetrics.forEach( i => {
-            (parseInt(i.race) === fieldMapping['americanIndian']) ? 
-                raceObject['americanIndian'] = parseInt(i.raceCount)
-            :
-            (parseInt(i.race) === fieldMapping['asian']) ?
-                raceObject['asian'] = parseInt(i.raceCount)
-            :
-            (parseInt(i.race) === fieldMapping['africanAmerican']) ?
-                raceObject['africanAmerican'] = parseInt(i.raceCount)
-            :
-            (parseInt(i.race) === fieldMapping['latino']) ?
-                raceObject['latino'] = parseInt(i.raceCount)
-            :
-            (parseInt(i.race) === fieldMapping['middleEastern']) ?
-                raceObject['middleEastern'] = parseInt(i.raceCount)
-            :
-            (parseInt(i.race) === fieldMapping['nativeHawaiian']) ?
-                raceObject['nativeHawaiian'] = parseInt(i.raceCount)
-            :
-            (parseInt(i.race) === fieldMapping['white']) ?
-                raceObject['white'] = parseInt(i.raceCount)
-            :
-            (parseInt(i.race) === fieldMapping['none']) ?
-                raceObject['none'] = parseInt(i.raceCount)
-            :
-                raceObject['other'] = parseInt(i.raceCount)
-            return raceObject;
-            
-    })}
-
-const filterAgeMetrics = (participantsAgeMetrics, ageRange) => {
-    participantsAgeMetrics.forEach( i => {
-        let participantYear = humanReadableY() - parseInt(i.birthYear)
-        sortAgeRange(participantYear, ageRange)
-        return ageRange;
-        }
-    )
-}
-
-const sortAgeRange = (participantYear, ageRange) => {
-    
-        (participantYear >= 40 && participantYear <= 45) ? 
-                ageRange['40-45']++
-        : 
-        (participantYear >= 46 && participantYear <= 50) ?
-                ageRange['46-50']++
-        : 
-        (participantYear >= 51 && participantYear <= 55) ?
-                ageRange['51-55']++
-        :
-        (participantYear >= 56 && participantYear <= 60) ?
-                ageRange['56-60']++
-        : 
-        (participantYear >= 61 && participantYear <= 65) ?
-                ageRange['61-65']++
-        : ""
-}
-
-const renderStackBarChart = (participantGenderResponse, participantRaceResponse, participantAgeRangeResponse, id) => {
-
-    const participantGenderResponseF = participantGenderResponse.female;
-    const participantGenderResponseM = participantGenderResponse.male;
-    const participantGenderResponseI = participantGenderResponse.intersex;
-    const participantGenderResponseU = participantGenderResponse.unknown;
- 
-    const totalGenderResponse = participantGenderResponseF + participantGenderResponseM + participantGenderResponseI + participantGenderResponseU
-
-    const participantRaceResponseAI = participantRaceResponse.americanIndian;
-    const participantRaceResponseA = participantRaceResponse.asian;
-    const participantRaceResponseAA = participantRaceResponse.africanAmerican;
-    const participantRaceResponseL = participantRaceResponse.latino;
-    const participantRaceResponseME = participantRaceResponse.middleEastern;
-    const participantRaceResponseNH = participantRaceResponse.nativeHawaiian;
-    const participantRaceResponseW = participantRaceResponse.white;
-    const participantRaceResponseO = participantRaceResponse.none;
-    const participantRaceResponseU = participantRaceResponse.other;
-        
-    const totalRaceResponse = participantRaceResponseAI + participantRaceResponseA + participantRaceResponseAA +
-                                participantRaceResponseL + participantRaceResponseME + participantRaceResponseNH +
-                            participantRaceResponseW + participantRaceResponseO + participantRaceResponseU
-                            
-
-    const participantAgeRangeResponse1 = participantAgeRangeResponse['40-45'];
-    const participantAgeRangeResponse2 = participantAgeRangeResponse['46-50'];
-    const participantAgeRangeResponse3 = participantAgeRangeResponse['51-55'];
-    const participantAgeRangeResponse4 = participantAgeRangeResponse['56-60'];
-    const participantAgeRangeResponse5 = participantAgeRangeResponse['61-65'];
-
-    const totalAgeRangeResponse = participantAgeRangeResponse1 + participantAgeRangeResponse2 + participantAgeRangeResponse3 + participantAgeRangeResponse4 + participantAgeRangeResponse5
-    const totalVerifiedParticipants = totalGenderResponse + totalRaceResponse + totalAgeRangeResponse
-
-    const genderPercent = Math.round((totalGenderResponse / totalVerifiedParticipants) * 100)
-    const racePercent = Math.round((totalRaceResponse / totalVerifiedParticipants) * 100)
-    const ageRangePercent = Math.round((totalAgeRangeResponse / totalVerifiedParticipants) * 100)
-
-
-    let ageTrace1 = {
-        y: ['Age'],
-        x: [participantAgeRangeResponse1],
-        type: 'bar',
-        name: '40-45',
-        text: `Total: ${ageRangePercent}%`,
-        orientation: 'h'
-    };
-
-    let ageTrace11 = {
-        y: ['Age'],
-        x: [participantAgeRangeResponse2],
-        type: 'bar',
-        name: '46-50',
-        text: `Total: ${ageRangePercent}%`,
-        orientation: 'h'
-    };
-
-    let ageTrace12 = {
-        y: ['Age'],
-        x: [participantAgeRangeResponse3],
-        type: 'bar',
-        name: '51-55',
-        text: `Total: ${ageRangePercent}%`,
-        orientation: 'h'
-    };
-
-    let ageTrace13 = {
-        y: ['Age'],
-        x: [participantAgeRangeResponse4],
-        type: 'bar',
-        name: '56-60',
-        text: `Total: ${ageRangePercent}%`,
-        orientation: 'h'
-    };
-
-    let ageTrace14 = {
-        y: ['Age'],
-        x: [participantAgeRangeResponse5],
-        type: 'bar',
-        name: '61-66',
-        text: `Total: ${ageRangePercent}%`,
-        orientation: 'h'
-    };
-
-    let raceTrace = {
-        y: ['Race Binary'],
-        x: [participantRaceResponseAI],
-        xaxis: 'x2',
-        yaxis: 'y2',
-        type: 'bar',
-        name: 'American Indian',
-        text: `Total: ${racePercent}%`,
-        orientation: 'h'
-    };
-    let raceTrace1 = {
-        y: ['Race Binary'],
-        x: [participantRaceResponseA],
-        xaxis: 'x2',
-        yaxis: 'y2',
-        type: 'bar',
-        name: 'Asian',
-        text: `Total: ${racePercent}%`,
-        orientation: 'h'
-    };
-    let raceTrace2 = {
-        y: ['Race Binary'],
-        x: [participantRaceResponseAA],
-        xaxis: 'x2',
-        yaxis: 'y2',
-        type: 'bar',
-        name: 'African American',
-        text: `Total: ${racePercent}%`,
-        orientation: 'h'
-    };
-    let raceTrace3 = {
-        y: ['Race Binary'],
-        x: [participantRaceResponseL],
-        xaxis: 'x2',
-        yaxis: 'y2',
-        type: 'bar',
-        name: 'Latino',
-        text: `Total: ${racePercent}%`,
-        orientation: 'h'
-    };
-    let raceTrace4 = {
-        y: ['Race Binary'],
-        x: [participantRaceResponseME],
-        xaxis: 'x2',
-        yaxis: 'y2',
-        type: 'bar',
-        name: 'Middle Eastern',
-        text: `Total: ${racePercent}%`,
-        orientation: 'h'
-    };
-    let raceTrace5 = {
-        y: ['Race Binary'],
-        x: [participantRaceResponseNH],
-        xaxis: 'x2',
-        yaxis: 'y2',
-        type: 'bar',
-        name: 'Native Hawaiian',
-        text: `Total: ${racePercent}%`,
-        orientation: 'h'
-    };
-
-
-    let raceTrace6 = {
-        y: ['Race Binary'],
-        x: [participantRaceResponseW],
-        xaxis: 'x2',
-        yaxis: 'y2',
-        type: 'bar',
-        name: 'White',
-        text: `Total: ${racePercent}%`,
-        orientation: 'h'
-    };
-
-    let raceTrace7 = {
-        y: ['Race Binary'],
-        x: [participantRaceResponseO],
-        xaxis: 'x2',
-        yaxis: 'y2',
-        type: 'bar',
-        name: 'Other',
-        text: `Total: ${racePercent}%`,
-        orientation: 'h'
-    };
-    let raceTrace8 = {
-        y: ['Race Binary'],
-        x: [participantRaceResponseU],
-        xaxis: 'x2',
-        yaxis: 'y2',
-        type: 'bar',
-        name: 'Unknown',
-        text: `Total: ${racePercent}%`,
-        orientation: 'h'
-    };
-
-    let genderTrace = {
-        y: ['Sex'],
-        x: [participantGenderResponseF],
-        xaxis: 'x3',
-        yaxis: 'y3',
-        type: 'bar',
-        name: 'Female',
-        text: `Total: ${genderPercent}%`,
-        orientation: 'h'
-    };
-
-    let genderTrace2 = {
-        y: ['Sex'],
-        x: [participantGenderResponseM],
-        xaxis: 'x3',
-        yaxis: 'y3',
-        type: 'bar',
-        name: 'Male',
-        text: `Total: ${genderPercent}%`,
-        orientation: 'h'
-    };
-
-    let genderTrace3 = {
-        y: ['Sex'],
-        x: [participantGenderResponseI],
-        xaxis: 'x3',
-        yaxis: 'y3',
-        type: 'bar',
-        name: 'Intersex',
-        text: `Total: ${genderPercent}%`,
-        orientation: 'h'
-    };
-
-    let genderTrace4 = {
-        y: ['Sex'],
-        x: [participantGenderResponseU],
-        xaxis: 'x3',
-        yaxis: 'y3',
-        type: 'bar',
-        name: 'Unknown',
-        text: `Total: ${genderPercent}%`,
-        orientation: 'h'
-    };
-
-    let data = [
-        ageTrace1, ageTrace11, ageTrace12, ageTrace13, ageTrace14,
-        raceTrace, raceTrace1, raceTrace2, raceTrace3, raceTrace4, raceTrace5, raceTrace6, raceTrace7, raceTrace8,
-        genderTrace, genderTrace2, genderTrace3, genderTrace4
-    ];
-
-
-    let layout = {
-        barmode: 'stack',
-        showlegend: false,
-        paper_bgcolor: 'rgba(0,0,0,0)',
-        plot_bgcolor: 'rgba(0,0,0,0)',
-
-
-        // colorway: ['#111E6C', '#1D2951', '#0E4D92', '#0080FF', '#008ECC', '#6593F5'],  
-        autosize: false,
-        width: 600,
-        height: 300,
-        title: 'Demographics of Verified Participants',
-        yaxis: { domain: [0, 0.266], automargin: true },
-        xaxis: { showgrid: false, automargin: true, showticklabels: false, title: { text: `Total number of verified participants: ${totalVerifiedParticipants}` } },
-        xaxis3: { anchor: 'y3', showticklabels: false, automargin: true },
-        xaxis2: { anchor: 'y2', showticklabels: false, automargin: true },
-        yaxis2: { domain: [0.366, 0.633], automargin: true },
-        yaxis3: { domain: [0.733, 1], automargin: true },
-
-    };
-
-    Plotly.newPlot(id, data, layout, { responsive: true, displayModeBar: false });
-}
-
-
-// const renderSiteSelection = (data) => {
-//     let template = `<label class="col-md-1 col-form-label" for="selectSites">Select IHCS</label><div class="col-md-4 form-group"><select id="selectSites" class="form-control">`
-//     template += `<option value="all">All</option>`
-//     data.map(dt => dt['siteCode']).forEach(siteCode => {
-//         template += `<option value="${siteCode}">${JSON.parse(localStorage.conceptIdMapping)[siteCode] && [siteCode]['Variable Name'] || [siteCode]['Variable Label']}</option>`
-//     })
-//     template += '</select></div>'
-//     document.getElementById('siteSelection').innerHTML = template;
-//     addEventSiteSelection(data);
-// }
-
-// const addEventSiteSelection = (data) => {
-//     const select = document.getElementById('selectSites');
-//     select.addEventListener('change', () => {
-//         let filteredData = '';
-//         if (select.value === 'all') filteredData = data;
-//         else filteredData = data.filter(dt => dt['siteCode'] === parseInt(select.value));
-//         renderAllCharts(filteredData);
-//     })
-// }
 
 
 const clearLocalStroage = () => {
@@ -1282,7 +567,15 @@ const clearLocalStroage = () => {
     animation(false);
     delete localStorage.dashboard;
     window.location.hash = '#';
+}
 
+const filterDatabySiteCode = (resultHolder, filteredResults, siteKeyFilter) => {
+    filteredResults.filter(i => {
+    if(i.siteCode === siteKeyFilter) {
+        resultHolder.push(i)
+    }
+    return resultHolder;
+    });    
 }
 
 const renderParticipantsNotVerified = async () => {
@@ -1415,10 +708,6 @@ const renderParticipantsAll = async () => {
         window.location.hash = '#';
     }
 }
-
-
-
-
 
 const eventNotVerifiedButton = (siteKey) => {
     const notVerifiedBtns = document.getElementsByClassName('participantNotVerified');
