+++ siteManagerDashboard/siteKeysToName.js
@@ -1,0 +1,31 @@
+
+
+export const keyToNameObj = 
+{
+
+    452412599 : "Kaiser Permanente Northwest",
+    531629870 : "HealthPartners",
+    657167265 : "Sanford Health",
+    548392715 : "Henry Ford Health System",
+    303349821 : "Marshfield Clinic",
+    125001209 : "Kaiser Permanente Colorado",
+    809703864 : "University of Chicago Medicine",
+    13 : "National Cancer Institute",
+    300267574 : "Kaiser Permanente Hawaii",
+    327912200 : "Kaiser Permanente Georgia"
+
+}
+
+export const nameToKeyObj = 
+{
+    "kpNW": 452412599,
+    "hPartners" : 531629870,
+    "snfrdHealth": 657167265,
+    "hfHealth": 548392715,
+    "maClinic": 303349821,
+    "kpCO": 125001209,
+    "uChiM": 809703864,
+    "nci": 13,
+    "kpHI": 300267574,
+    "kpGA": 327912200 
+}
